refactor: Move stopAllProcessesOnExit to processLifecycle.ts
